refactor: DTO와 인기 테마 패키지 분리
