+++ src/main/java/roomescape/controller/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 import roomescape.domain.ReservationTime;
