+++ src/main/java/roomescape/controller/user/ReservationStatusController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.dto.ReservationStatusResponse;
+import roomescape.controller.dto.response.ReservationStatusResponse;
 import roomescape.service.ReservationLookupService;
 
 import java.util.List;
