+++ src/main/java/roomescape/controller/user/ReservationWaitingController.java
@@ -6,8 +6,8 @@ import jakarta.validation.constraints.Positive;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import roomescape.controller.dto.ReservationWaitingRequest;
-import roomescape.controller.dto.ReservationWaitingResponse;
+import roomescape.controller.dto.request.ReservationWaitingRequest;
+import roomescape.controller.dto.response.ReservationWaitingResponse;
 import roomescape.domain.WaitingWithTurn;
 import roomescape.service.ReservationWaitingService;
