+++ src/main/java/roomescape/domain/populartheme/PopularThemeCondition.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.populartheme;
 
 import java.time.LocalDate;
