+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.resource.NoResourceFoundException;
-import roomescape.controller.dto.ErrorResponse;
+import roomescape.controller.dto.response.ErrorResponse;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
