+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,10 +3,10 @@ package roomescape.service;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.PopularTheme;
-import roomescape.domain.PopularThemeCondition;
-import roomescape.domain.PopularThemePolicy;
 import roomescape.domain.Theme;
+import roomescape.domain.populartheme.PopularTheme;
+import roomescape.domain.populartheme.PopularThemeCondition;
+import roomescape.domain.populartheme.PopularThemePolicy;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.RoomescapeException;
 import roomescape.repository.ReservationRepository;
