+++ src/test/java/roomescape/domain/populartheme/PopularThemeConditionTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.populartheme;
 
 import org.junit.jupiter.api.Test;
